refactor: 프로젝트 상세 페이지 hooks 폴더를 colocation 방식으로 수정 및 컴포너트 폴더 분리
