+++ frontend/src/pages/project-detail/Carousel/Carousel.styled.ts
