+++ frontend/src/pages/project-detail/Carousel/Carousel.tsx
