+++ frontend/src/pages/project-detail/Carousel/hooks/useSlide.ts
