+++ frontend/src/pages/project-detail/OverviewSection/OverviewSection.styled.ts
