+++ frontend/src/pages/project-detail/TechStacksSection/TechStacksSection.styled.ts
