+++ frontend/src/pages/project-detail/TitleSection/TitleSection.styled.ts
