+++ frontend/src/pages/project-detail/components/SectionTitle/SectionTitle.styled.ts
