+++ frontend/src/pages/project-detail/hooks/useProjectDetailData.ts
